+++ src/main/java/com/mymovie/service/MovieService.java
@@ -48,5 +48,5 @@ public interface MovieService {
      * @param movieId
      * @return
      */
-    Movie findMovieById(int movieId);
+    Movie findMovieById(long movieId);
 }
